requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
openai==0.28.1
requests==2.31.0
python-dotenv==1.0.0
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
python-dateutil==2.8.2
